+++ js/components/show-big-picture.js
@@ -1,34 +1,11 @@
-const body = document.querySelector('body');
-const bigPicture = document.querySelector('.big-picture');
-const commentCount = bigPicture.querySelector('.social__comment-count');
-const commentList = bigPicture.querySelector('.social__comments');
-const commentsLoader = bigPicture.querySelector('.comments-loader');
+import {
+  renderComments
+} from '../components/render-сomments';
+
+
+const body = document.body;
+const bigPicture = body.querySelector('.big-picture');
 const cancelButton = bigPicture.querySelector('.big-picture__cancel');
-
-const createComment = ({ url, name, message }) => {
-  const comment = document.createElement('li');
-  comment.innerHTML =
-    '<img class="social__picture" src="" alt="" width="35" height="35"><p class="social__text"></p>';
-  comment.classList.add('social__comment');
-  const image = comment.querySelector('.social__picture');
-  image.src = url;
-  image.alt = name;
-  comment.querySelector('.social__text').textContent = message;
-
-  return comment;
-};
-
-const renderComments = (comments) => {
-  commentList.innerHTML = '';
-
-  const fragment = document.createDocumentFragment();
-  comments.forEach((comment) => {
-    const commentElement = createComment(comment);
-    fragment.append(commentElement);
-  });
-
-  commentList.append(fragment);
-};
 
 const hideBigPicture = () => {
   bigPicture.classList.add('hidden');
@@ -47,23 +24,23 @@ const onCancelButtonClick = () => {
   hideBigPicture();
 };
 
-const renderPictureDetails = ({ url, likes, description }) => {
+
+const showBigPicture = ({ url, likes, description, comments }) => {
   const img = bigPicture.querySelector('.big-picture__img img');
+  const likesCount = bigPicture.querySelector('.likes-count');
+  const socialCaption = bigPicture.querySelector('.social__caption');
+
   img.src = url;
   img.alt = description;
-  bigPicture.querySelector('.likes-count').textContent = likes;
-  bigPicture.querySelector('.social__caption').textContent = description;
-};
+  likesCount.textContent = likes;
+  socialCaption.textContent = description;
 
-const showBigPicture = (data) => {
   bigPicture.classList.remove('hidden');
   body.classList.add('modal-open');
-  commentsLoader.classList.add('hidden');
-  commentCount.classList.add('hidden');
-  document.addEventListener('keydown', onEscKeyDown);
 
-  renderPictureDetails(data);
-  renderComments(data.comments);
+  renderComments(comments, bigPicture);
+
+  document.addEventListener('keydown', onEscKeyDown);
 };
 
 cancelButton.addEventListener('click', onCancelButtonClick);
